# openapi.yaml
openapi: 3.0.1
info:
  title: Steam Plugin
  description: A plugin that allows to search for games and see information about a game.
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /game/{id}:
    get:
      operationId: getGame
      summary: Returns information about a game
      parameters:
      - in: path
        name: id
        schema:
            type: integer
        required: true
        description: The steam game id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/game'
...
components:
  schemas:
    game:
      type: object
      properties:
        steam_appid:
          type: integer
          description: The ID of the game on Steam.
        name:
          type: string
          description: The name of the game.
        developers:
          type: array
          items:
            type: string
          description: The developers of the game.
        publishers:
          type: array
          items:
            type: string
          description: The publishers of the game.
        categories:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              description:
                type: string
          description: Categories of the game.
        pc_requirements:
          type: object
          properties:
            minimum:
              type: string
            recommended:
              type: string
          description: PC requirements for the game.
        mac_requirements:
          type: object
          properties:
            minimum:
              type: string
            recommended:
              type: string
          description: PC requirements for the game.
        linux_requirements:
          type: object
          properties:
            minimum:
              type: string
            recommended:
              type: string
          description: PC requirements for the game.
        website:
          type: string
          description: The website of the game.
        required_age:
          type: string
          description: The required age to play the game.
        release_date:
          type: object
          properties:
            coming_soon:
              type: boolean
            date:
              type: string
          description: The release date of the game.
        platforms:
          type: object
          properties:
            windows:
              type: boolean
            mac:
              type: boolean
            linux:
              type: boolean
          description: The platforms the game is available on.
        genres:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              description:
                type: string
          description: Genres of the game.
        recommendations:
          type: object
          properties:
            total:
              type: integer
          description: Total number of recommendations.
        achievements:
          type: object
          properties:
            total:
              type: integer
            highlighted:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
          description: Achievements related to the game.
        dlc:
          type: array
          items:
            type: integer
            description: Steam DLC game id
      required:
        - steam_appid
        - name
        - developers
        - publishers
        - platforms
        - genres
        - recommendations
        - achievements
...